Initial implementation of the Chaos Visual Debugger plugin
This CL contains the plugin which adds the Window tool for visualizing recorded physics simulations and an engine module that defines the data format (although for this first submission is just a placeholder).

The recording implementation is not included as it is pretty much a placeholder like the data format itself and will change shortly.

The tool can be open from the Menu Tools->Debug->Chaos Visual Debugger

Included functionality:

- 3D Viewport to see the recorded elements of the simulation (Independent of the Editor). The current placeholder data format only contains the Position/Rotation of particles so the visualization is limited to that for now.
- Simple playback controls
- Instances of World Outliner, Details panel, and Output log
- Ability to load and reproduce cvd files (format defined in this same CL)

Current issues and limitations:
- Outliner/Viewport and details panel are still using global selection events so what is selected in the tool also affects the Editor. This will be addressed in the next commits when we implement derived versions of these elements and change to a local selection event.
- All the UI icons are placeholders, copies from base engine icons (some slightly modified).
- Settings button in the UI is not implemented

Everything is pretty much in flux. The next focus and likely next CL will be one dedicated to implementing a proper first version of the recording data format + the recording logic and after that refocus on the visualization tool again. Also, it is likely that the data format types will be moved into the Chaos module to be able o re-use existing types.

#rb benn.gallagher
#preflight 640feb13c41a0a2a77f27237

[CL 24701553 by sergio gardeazabal in ue5-main branch]