[Backout] - CL24692954 #fyi dave.jones2 Original CL Desc ----------------------------------------------------------------- UE-97716 - Private and Protected functions do not display in the context menu in child BPs
We need the classes used by the target function and the graph to be consistent. Specifically, we could run into scenarios where the TargetFunction would be the skeleton class and *not* the generated class. This would cause subsequent IsChildOf usage in CanFunctionBeUsedInGraph to fail, since they're technically two different classes.

In this case, we can explicitly use the skeleton classes from both to remove any ambiguity.

#jira UE-97716
#preflight 64149a73ca2afe3ee6a48565
#rb phillip.kavan

[CL 24719137 by dave jones2 in ue5-main branch]