Moved rendering the debug draw helper execution to after post process pass to avoid ghosting of TAA/TSR/MB.
The EditorPrimitives compositing code was the only PostProcess pass that composited appropriately, but was not setup to accept non-editor draws. DrawDebug also needs to be available to Debug and Development builds, so moving the debug draw calls to this pass directly was not straight forward, and there was no other existing option.

SimpleElement draws used by debug were also non-trivial to setup with velocity to compensate for Temporal/Blur passes, so fixing this in the base pass was also not straight forward.

Then an additional issue was that the color and depth textures used in post process after temporal passes were mismatched in size (as depth is not necessarily automatically included in the temporal upscale pass). However EditorPrimitives had hard coded temporal upsample depth code in it's main AddEditorPrimitivesPass function.

Refactored PostProcessEditorPrimitives code to move the hard coded temporal upscale depth pass to a common file which is protected by UE_ENABLE_DEBUG_DRAWING (which is blocked from compiling in shipping, and works with editor). Also ensured the relevant shader compiles remain protected by these macros. The temporal and populate functions have been slightly rewritten so that a generic temporal history is created rather than an editor only one, and also populate depth can now also output the background colour (to reduce unnecessary draw calls as, depending on the setup, the previous scene may still needs to be drawn to the output color texture).

After that refactor, there is a new set of draw debug composite files which now also use the temporal upsample depth code (when color and depth are mismatched, ignored if not), now that it has been isolated into it's own re-usable structure. This means that we can use a simple draw elements call, stripped of all shader variables except the view, to overlay the draw debug primitives on the scene color along with upscaled depth.

Also moved draw debug to it's own SimpleElement array in the View to avoid conflating it with other simple element draws (e.g. non-opaque, or ones setup in plugins). This code can be extended for those at a later date if needed though.

#jira UE-177222
#rb Jason.Nadro, Massimo.Tristano
#preflight 64189a2a3f3d31c94aa0c48d
#fyi Guillaume.Abadie

[CL 24731449 by Jon Cain in ue5-main branch]