Removed the EditorStorage world type.
The EditorStorage world type was introduced as MASS had a dependency on  having a world, but none of the existing worlds could be made to work. This limitation has since been removed from MASS which means there's no current use for EditorStorage. To avoid adding additional maintenance and since EditorStorage was introduce within 5.2, it's now being removed before release. If a use case is found at a later time, this CL can be used to reintroduce EditorStorage.

#jira UE-180332
#rb Mieszko.Zielinski
#preflight 6419d4bb13231a282690c8b4
#lockdown jeanmichel.dignard

[CL 24739070 by ronald koppers in ue5-main branch]