Chaos Cloth Dataflow: Add StaticMeshImportNode.
This will can be used to import a static mesh to start a new cloth asset. Optionally use the UVs to define the pattern data.

Internally using conversion from Mesh Description to FDynamicMesh3, and then importing the FDynamicMesh3. The existing FCollectionClothLodFacade::Initialize has been updated to also internally create an FDynamicMesh3 so the code can be shared. There were various bugs related to pattern edge and seam generation in the previous Initialize code (in particular it was possible for bowtie vertices to not be seamed back together, and no seams were generated between separate islands that were created as part of the unwrap process). Hopefully they've been fixed.

Of note, FDynamicMesh3 does not allow non-manifold edges. Edges are duplicated in the FDynamicMesh3 in the conversion process if a non-manifold edge is found. This split is reflected as a new Seam in the ClothAsset.

#preflight 641c9837b72410fc17289d41
#rb tyson.brochu kriss.gossart

[CL 24765953 by Alex McAdams in ue5-main branch]