Horde: Remove AgentCollection.TryUpdateAsyncWithRetries() (added in CL 20241607).
Operations on agent documents are meant to be compare-and-swap operations, hence the dependency on the UpdateIndex matching. Applying the same update to an updated document is not guaranteed to be valid, and can result in state being incorrectly overwritten. Update errors must to be propagated back to the caller to determine the correct course of action for an updated agent document.

#preflight none
#fyi Carl.Bystrom

[CL 24811764 by Ben Marsh in ue5-main branch]