Path Tracer: Fix incorrect pdfs returned by sample method
When refactoring the code for splitting diffuse from specular, a subtle error was introduced that could cause materials to render incorrectly under certain illumination conditions. This could only be seen when comparing the direct illumination output under MISMode 0,1,2. While modes 0 and 1 (corresponding to BSDF and Light sampling respetively) matched in all cases, the MIS combination of the two did not and frequently rendered darker.

The reason is that the eval function uses MIS across all lobes, while sample only returns the result of a single lobe. The weight would lead to the correct result, but the pdf did not compare correctly with the one returned by eval, leading to incorrect results when MIS favored BSDF sampling over light sampling.

To fix this, all multi-lobed shading models need to account for the probability of generating the same direction through all other lobes.

Along the way, change to extinction encoding for solid glass to avoid precision issues with transmission color in solidglass.

Also apply the RescaleRandomNumber function consistently across material sampling routines.


#jira UE-181325
#rb Aleksander.Netzel
#preflight 64220959974dfaa53cb614c8

[CL 24814539 by chris kulla in ue5-main branch]