UE-97716 - Private and Protected functions do not display in the context menu in child BPs (resubmit)
We need the classes used by the target function and the graph to be consistent. Specifically, we could run into scenarios where the TargetFunction would be the skeleton class and *not* the generated class. This would cause subsequent IsChildOf usage in CanFunctionBeUsedInGraph to fail, since they're technically two different classes.

In this case, if the target function comes from a BP, we can explicitly use the skeleton classes from both to remove any ambiguity.

#jira UE-97716
#preflight 64234b57974dfaa53c0e2a89
#rb phillip.kavan

[CL 24826011 by dave jones2 in ue5-main branch]