Sequencer: user-defined bindings
This change adds the ability to define custom logic (via the directory blueprint) for object bindings. Instead of binding to the usual possessable or spawnable, this will optionally run arbitrary BP logic defined by the user to resolve the bound object. This makes it possible, for instance, to bind to the player pawn, or any other actor that might not exist at edit time.

The user interface is similar to that of the event track, and therefore reuses the director blueprint endpoint picker. There is however a new blueprint extension to compile these functions into the spawnables/possessables, and of course a spot to call these functions when appropriate.

#jira UE-170204
#rb andrew.rodham
#preflight 6425f08c861bf344ced97d58

[CL 24861303 by ludovic chabant in ue5-main branch]