Sequencer: Fixed a crash caused by unloading a spawnable level instance
Destroying a spawnable level instance will cause a GC to run, which in turn will cause the sequencer ECS to tag garbage. Unfortunately this is invalid because the ECS is currently being iterated so is locked for mutation. The fix is to cache all the destroyed spawnables into an array on the stack, then destroy them outside of an iteration.

#rb
#jira UE-181788
#preflight 6426d5c99834526476998961

[CL 24869867 by Andrew Rodham in ue5-main branch]