Path Tracer: Speedup OpenImageDenoise plugin
Re-use memory allocations for CPU buffers in OpenImageDenoise plugin. Instead of creating a new array every time we run the denoiser, try to keep the old one around, assuming that the resolution is rarely changing. Also try to re-use the oidn filter as long as the pixel array and resolution has not changed.

Change from ReadSurfaceData method which does its own memory allocations to directly using LockTexture2D which lets us read the GPU texture data directly.

This saves up to 50ms of denoising time for typical editor resolutions.

Upgrade from OpenImageDenoise 1.4.0 to 1.4.1 as this fixes a crash with in-place denoising which we are now using. I did not update to the latest 1.4.3 to avoid changing the tbb dependency.

#jira none
#rb Tiantian.Xie
#preflight 642f6928ce7fbea1eebed616

[CL 24969720 by chris kulla in ue5-main branch]