Sequencer Playlists: Partially restore prior approach to initiating holds.
Reliably mapping between state in the pending take to the equivalent in the destination take turns out to be difficult due to some missing lifecycle events in Take Recorder. This is a conservative fix that ensures the UI behaves correctly after Take Recorder has started. The tradeoff is that, as in previous releases, the effect on the scene of any sequence held prior to beginning recording will not be visible until recording begins.

#jira UE-180065
#rb jason.walter
#preflight 643497b3b306e98c634a6072

[CL 25002668 by zach brockway in ue5-main branch]