Fix for editor viewports not drawing post process scene color and depth correctly when both editor primitives/selection outline is enabled at the same time as debug draws.
Debug primitive draws will only output the debug color and depth values onto the viewport location given due to SimpleElementShaders being locked into a specific draw setup. As a result, if the viewport MinX/Y value is not 0, then the color and depth textures must be appropriately drawn out in an alternative way to the target render targets for the debug primitive composite to match the location in the viewport. This can likely be achieved with multiple calls to AddDrawTexturePass, but instead this change alters the PopulateDepthPass to allow for an alternative output format which draws the color and depth to the required viewport location.

For future work, it is likely the best way forward would be to rethink how debug primitives are drawn, as the SimpleElementShaders and their setup are locked in and difficult to alter. For now this is likely the best way to compensate until debug draws can be refactored entirely.

Also removed a redundant argument in TemporalUpscaleDepthPass.

#rb Jason.Nadro
#jira none
#preflight 6437fa441f33a0b34a7f4d78

[CL 25024370 by Jon Cain in ue5-main branch]