The Paper2D Atlas will now support converting the source data of other texture to BGRA8 when creating the atlas.
Note the atlas still only support BGRA8 for the source data of the texture build. This only make so that it won't leave an empty space when a sprite is using a non BGRA8 texture. Also the atlas are still experimental and not supported so this change is only to unblock the users from a recent change to the texture import.

#jira UE-182530
#preflight 643452194c3ccbbdf1805b9d
#rb JeanLuc.Conrenthin

[CL 25024498 by Julien StJean in ue5-main branch]