Further optimization for moving actor with many static mesh components.  Only invalidate the static lighting cache at the start of a move operation.  We don't invalidate the cache in "GizmoManipulationStarted", because the user can click on a handle (calling the start function) without actually moving, which should have no effect.  So a flag is used to detect the first time "GizmoManipulationDeltaUpdate" is called, and subsequent calls can skip repeatedly invalidating the lighting cache.
With this fix, on my machine, the cost of moving a blueprint with 1000 items drops from 45 ms to 7 ms.  Baseline tick is 8 ms (120 fps) when idling, so the results while moving improve from 19 fps to 65 fps.  The remaining time is mostly split between physics and navigation update.  Invalidating the static lighting cache is so costly because it both invalidates the render state of every item, and notifies packages of the change.

#jira UE-147154
#rnx
#rb jamie.dale
#preflight 63ff4748e32cdb7d9aacab16

[CL 25050136 by jason hoerner in ue5-main branch]