Path Tracer: Add decal support with Substrate enabled
Refactor the decal apply logic so that we adjust diffuse/specular _after_ having applied the decals so that we can properly handle BaseColor even in the presence of subsurface scattering.

Refactor EnableDBuffer logic to remove the decal data after the second trace so we don't need to check it again. This keeps the dbuffer logic more self contained.

PT_Decals in EngineTest is mostly passing, but there are a few differences that need to be investigated more carefully. I suspect that some of them might be due to the round trip between diffuse/specular color and metallic/specular workflow.

#rb Sebastien.Hillaire,Tiago.Costa
#jira UE-173840
#preflight 643eea360a5a4b944e936ffc

[CL 25117919 by chris kulla in ue5-main branch]