UVEditor: Enable support for a new UV distortion visualization display setting.
When enabled, it computes a per triangle distortion value of the triangle's UVs vs it's world space positions. This is then displayed as a color blind friendly color overlay on the Unwrap mesh. The overlay is designed to update automatically whenever the preview is updated, supporting both tool operation edits and manual updates with the gizmo.

#rb semion.piskarev
#preflight 64419af64944ef7892eea920

[CL 25155790 by nathan mitchell in ue5-main branch]