USDImporter: add wrapping for UsdVariantSet/UsdVariantSets to enable directly authoring variant sets on a prim
As part of this change, all of FUsdPayloads, FUsdReferences, and FUsdVariantSets are
only constructable through the appropriate Get...() function on FUsdPrim, since those
classes are all only meaningful in the context of a particular prim. This also aligns the
UE-side wrapping more closely with the USD API.

In addition, the instantiations of FScopedUnrealAllocs objects were removed from
all member function implementations. This avoids "forcing" any particular allocator
from being used when constructing objects, allowing the calling code to provide a
context using either of the UE or USD allocators.

Assignment and conversion operators were added for FUsdPayloads and FUsdReferences
as well, which were missing from their initial introduction.

#rb daniel.coelho
#preflight 64475b8727014596f3dc89d6

[CL 25185252 by matt johnson in ue5-main branch]