Chaos Caching: Correctly handle spawnable CacheManagers which are observing components that belong to the cache manager itself (i.e., geometry collections that are created by adding a CacheCollection to a scene).
The crux of the bug is that the SoftComponentRef.OtherActor will point to the spawnable template, not the spawned instance. This is the case for any spawnable observed component.

This fixes the particular case where the Component's owner is the CacheManager itself by setting OtherActor to nullptr, and passing in the CacheManager as the "OwningActor" at runtime. This includes an upgrade in Serialize to clear the OtherActor when it points to the CacheManager itself.

Tested this works with new spawnable cache managers as well as old existing sequences that were previously broken.

#jira UE-184250
#preflight 644851c11150e908d0d36fa8
#rb cedric.caillaud brice.criswell

[CL 25191788 by Alex McAdams in ue5-main branch]