Add PhysicsAssets to UEFN for Query/Probe only. All UI related to constraints and simulation has been hidden.
I updated a few of the main menus to use the new "Tool Menus" instead of the old "Menu Builder" so they could be hidden via the json permissions file. However, there are a bunch of contextual menus that I did not upgrade. Instead, I gated the creation of any simulation or constraint "legacy" menus via new properties in the PhysicsAssetEditorOptions which are set from the ValkyrieEditor. The Valkyrie Editor also changes the default create params (FPhysAssetCreateParams) to not create constraints, so constraints will not be created when creating new physics assets.

#preflight 644818dbfa6ba3cc5b5a818e
#rb rex.hill jason.stasik chris.caulfield

[CL 25192111 by alex mcadams in ue5-main branch]