Added logging an Error to UCustomizableObjectSystem::AddUncompiledCOWarning, and also added an optional parameter for the caller to add more info to the log (so we can easily log the source of the call to figure out what is being called that is trying to use a CO that hasn't been compiled yet).
Created a private helper function FCustomizableObjectInstanceDescriptor::AddUncompiledCOWarning(...), which makes sure the System exists, is valid, and isn't being destroyed before calling to AddUncompiledCOWarning on the System.

SetIntParameterSelectedOption(FString ParameterName,...) now ensures and calls to AddUncompiledCOWarning (which logs an error, makes an on-screen display, and sends a notfiication in Editor) if the CO that the descriptor is associated with is not compiled.  If it is compiled, but the named parameter is not found, it ensures (in non-shipping builds only) and logs an error (in all builds).

SetIntParameterSelectedOption((const int32 IntParamIndex, ...) now also ensure and calls AddUncompiledCOWarning  as the string version does if its CO is not compiled.  If it is compiled, but the IntParamIndex is invalid, it logs an error and bails.

(We should make similar ensure/logging changes to all other parameter type setters.
 I'll investigate that in the near future.)

#RB Joel.Anderson, Pere.Rifa



#UE5

[CL 25219708 by daniel broder in ue5-main branch]