Add cloth visualization to new cloth asset editor. This is just bringing over a copy of the existing debug draw features for skeletal mesh cloth. We expect to iterate and improve on this in the future beyond what is available in the legacy editor.
ClothEditorSimulationVisualization is a copy of ChaosSimulationEditorExtender, but with interface updates to match our current use (e.g., IClothingSimulation -> UChaosClothComponent). The Visualization is owned by the FChaosClothAssetEditorToolkit. A WeakPtr is shared with FChaosClothAssetEditor3DViewportClient.

Additional changes to make this work:
Added ability to disable and enable simulation
Added an InvalidateViews call when ticking the ToolKit. This ensures the views update while slate is throttling (e.g., when menus are open). This is what the other persona viewports do, so I thought it was OK?

#preflight 64518008de576521224a1005
#rb tyson.brochu kriss.gossart

[CL 25318845 by Alex McAdams in ue5-main branch]