Path Tracing: Implement refraction for Substrate materials
When a material is flagged to used refraction, compute the IOR and set it on the slab (for the bottom slabs only). For now this case is split into its own logic with

A different implementation is used for Thin vs Solid cases for now.

This implementation is still a work in progress. In particular, the treatment of the diffuse color (which controls the medium albedo) is not complete.

#rb none
#jira UE-173840
#preflight 6456d9ef994d8a3d0aa6d479

[CL 25366534 by chris kulla in ue5-main branch]