Fixed an initialization bug in TEDS for cooked builds.
The order-of-initialization for cooked editors appears to be different from standalone builds. As a result the static initialization trick used to impersonate Mass tags and fragments doesn't work. However since TEDS no only uses dynamic processors there's no longer a need to have the impersonation initialized before the CDOs are created (which would setup queries which in turn need tags and fragments).

#jira UE-185289
#rb jeanmichel.dignard
#preflight 645a7824013f3751586adf7d

[CL 25391253 by ronald koppers in ue5-main branch]