[Chaos VD] Adding support to recognize and load multiple solvers running at the same time on a ChaosVD recording.
Before this change, only the latest loaded solver data was shown in the UI to control.

Now, the tool provides a single timeline representation for Game Frames (Game thread ticks), and when you scrub that timeline, solvers are loaded/unloaded, and scrub to the closest "solver frame available" for that game frame.

A new tab was added that contains the old playback controls, but this time one for each available solver at the current game frame. These controls can be collapsed as desired.

The per-solver playback controls can be used to scrub in a more precise way to a specific solver frame, or solver step. When scrubbing a solver control, the game frame timeline gets automatically scrubbed to the correct game frame (these are not in sync, so doing it exactly is not possible).

All the approximate synchronization between timelines is done using the recorded platform cycles by using them as the lower bound for a binary search for the target "track spacetime". This is possible because Trace recordings and analysis are serial, so when we populate Game frame data and Solver frame data during Trace analysis, it is guaranteed to be populated in order based on the recorded platform cycles.

The API exposed in the ChaosVDRecording object to do these searches and conversion between "solver spacetimes" could be improved, but this is a starting point.

Not supported yet:

- Scrubbing on a solver timeline will update the game timeline for reference, but not other solvers' timelines, and therefore the scene data for these will not be updated on screen. But scrubbing the Game Frames timeline will bring everything in sync again.

[REVIEW]
#jira UE-181317
#preflight 645c02b2592448f08b2d52f4

[CL 25420199 by sergio gardeazabal in ue5-main branch]