Sequencer: Fix group manager's selection synchronization
When selection changes in the group manager tree, select the corresponding Sequencer items first. Then update the group manager tree's notion of what is selected. When this empties selection, the group item itself will be deselected. Any items within it will be synchronized with what is selected in Sequencer.

Also, fixed so this works with folder paths.

#jira UE-181607
#rb andrew.rodham
#preflight 645e92f95b775f11fbcf95e5

[CL 25460545 by max chen in ue5-main branch]